fix thread-safety
We now globally lock all the interpreters when we iterate over all the interpreters'
bump allocators to resolve which interpreter a `QualType` belongs to.
Similarly, we also lock all the interpreters to resolve the interpreter from a `Decl`.

I have also adopted most of the tests to run in parallel, hopefully detecting failures early.